Revert "App upgrade notification shown for incorrect app (#9806)"
This reverts commit eec4a46fdb9db248d5c393b56d0e9325c679bd5d.
