+++ .vscode/settings.json
@@ -50,7 +50,6 @@
     "storageclass",
     "testid",
     "tolerations",
-    "upgrader",
     "userpreferences",
     "virtualmachine",
     "vuex",
