--- shell/config/os.ts
@@ -1,2 +1,0 @@
-export const WINDOWS = 'windows';
-export const LINUX = 'linux';
