+++ shell/pages/c/_cluster/apps/charts/install.vue
@@ -36,7 +36,7 @@ import { ignoreVariables } from './install.helpers';
 import { findBy, insertAt } from '@shell/utils/array';
 import Vue from 'vue';
 import { saferDump } from '@shell/utils/create-yaml';
-import { LINUX, WINDOWS } from '@shell/config/os';
+import { LINUX, WINDOWS } from '@shell/store/catalog';
 
 const VALUES_STATE = {
   FORM: 'FORM',
