+++ shell/store/catalog.js
@@ -3,11 +3,14 @@ import { CATALOG as CATALOG_ANNOTATIONS } from '@shell/config/labels-annotations
 import { addParams } from '@shell/utils/url';
 import { allHash, allHashSettled } from '@shell/utils/promise';
 import { clone } from '@shell/utils/object';
-import { findBy, addObject, filterBy } from '@shell/utils/array';
+import { findBy, addObject, filterBy, isArray } from '@shell/utils/array';
 import { stringify } from '@shell/utils/error';
 import { classify } from '@shell/plugins/dashboard-store/classify';
 import { sortBy } from '@shell/utils/sort';
 import { importChart } from '@shell/utils/dynamic-importer';
+import { ensureRegex } from '@shell/utils/string';
+import { isPrerelease } from '@shell/utils/version';
+import difference from 'lodash/difference';
 import { lookup } from '@shell/plugins/dashboard-store/model-loader';
 
 const ALLOWED_CATEGORIES = [
@@ -28,6 +31,9 @@ const CERTIFIED_SORTS = {
   [CATALOG_ANNOTATIONS._PARTNER]:      2,
   other:                               3,
 };
+
+export const WINDOWS = 'windows';
+export const LINUX = 'linux';
 
 export const state = function() {
   return {
@@ -94,42 +100,19 @@ export const getters = {
     return sortBy(out, ['certifiedSort', 'repoName', 'chartName']);
   },
 
-  /**
-   * Returns matching chart for given parameters
-   * @returns
-   */
   chart(state, getters) {
     return ({
-      chart, // to extract parameters
-      key, // filtering mode 1
-      repoType, repoName, chartName, // filtering mode 2
-      preferRepoType, preferRepoName, // sorting
-      includeHidden
+      key, repoType, repoName, chartName, preferRepoType, preferRepoName, includeHidden
     }) => {
-      // Get data from the Chart
-      if ( chart && !key && !repoType && !repoName && !chartName) {
-        chartName = chart.metadata?.name;
-        preferRepoType = chart.metadata?.annotations?.[CATALOG_ANNOTATIONS.SOURCE_REPO_TYPE];
-        preferRepoName = chart.metadata?.annotations?.[CATALOG_ANNOTATIONS.SOURCE_REPO_NAME];
-        repoType = chart.repoType;
-        repoName = chart.repoName;
-      }
-
-      // Return nothing if cannot filter the charts, to avoid mismatches
-      if (!(key || (repoType && repoName && chartName))) {
-        return;
-      }
-
-      // Get data as key of the retrieved Charts (currently same as ID)
       if ( key && !repoType && !repoName && !chartName) {
-        const keyParts = key.split('/');
+        const parsed = parseKey(key);
 
-        repoType = keyParts[0];
-        repoName = keyParts[1];
-        chartName = keyParts[2];
+        repoType = parsed.repoType;
+        repoName = parsed.repoName;
+        chartName = parsed.chartName;
       }
 
-      let matchingCharts = filterBy(getters.charts, {
+      let matching = filterBy(getters.charts, {
         repoType,
         repoName,
         chartName,
@@ -137,19 +120,18 @@ export const getters = {
       });
 
       if ( includeHidden === false ) {
-        matchingCharts = matchingCharts.filter((x) => !x.hidden);
+        matching = matching.filter((x) => !x.hidden);
       }
 
-      if ( !matchingCharts.length ) {
+      if ( !matching.length ) {
         return;
       }
 
-      // Sorting by preference
       if ( preferRepoType && preferRepoName ) {
-        preferSameRepo(matchingCharts, preferRepoType, preferRepoName);
+        preferSameRepo(matching, preferRepoType, preferRepoName);
       }
 
-      return matchingCharts[0];
+      return matching[0];
     };
   },
 
@@ -477,6 +459,16 @@ export function generateKey(repoType, repoName, chartName) {
   return `${ repoType }/${ repoName }/${ chartName }`;
 }
 
+export function parseKey(key) {
+  const parts = key.split('/');
+
+  return {
+    repoType:  parts[0],
+    repoName:  parts[1],
+    chartName: parts[2],
+  };
+}
+
 function addChart(ctx, map, chart, repo) {
   const repoType = (repo.type === CATALOG.CLUSTER_REPO ? 'cluster' : 'namespace');
   const repoName = repo.metadata.name;
@@ -591,4 +583,87 @@ function filterCategories(categories) {
 
 function normalizeCategory(c) {
   return c.replace(/\s+/g, '').toLowerCase();
+}
+
+/*
+catalog.cattle.io/deplys-on-os: OS -> requires global.cattle.OS.enabled: true
+  default: nothing
+catalog.cattle.io/permits-os: OS -> will break on clusters containing nodes that are not OS
+  default if not found: catalog.cattle.io/permits-os: linux
+*/
+export function compatibleVersionsFor(chart, os, includePrerelease = true) {
+  const versions = chart.versions;
+
+  if (os && !isArray(os)) {
+    os = [os];
+  }
+
+  return versions.filter((ver) => {
+    const osPermitted = (ver?.annotations?.[CATALOG_ANNOTATIONS.PERMITTED_OS] || LINUX).split(',');
+
+    if ( !includePrerelease && isPrerelease(ver.version) ) {
+      return false;
+    }
+
+    if ( !os || difference(os, osPermitted).length === 0) {
+      return true;
+    }
+
+    return false;
+  });
+}
+
+export function filterAndArrangeCharts(charts, {
+  clusterProvider = '',
+  operatingSystems,
+  category,
+  searchQuery,
+  showDeprecated = false,
+  showHidden = false,
+  showPrerelease = true,
+  hideRepos = [],
+  showRepos = [],
+  showTypes = [],
+  hideTypes = [],
+} = {}) {
+  const out = charts.filter((c) => {
+    if (
+      ( c.deprecated && !showDeprecated ) ||
+      ( c.hidden && !showHidden ) ||
+      ( hideRepos?.length && hideRepos.includes(c.repoKey) ) ||
+      ( showRepos?.length && !showRepos.includes(c.repoKey) ) ||
+      ( hideTypes?.length && hideTypes.includes(c.chartType) ) ||
+      ( showTypes?.length && !showTypes.includes(c.chartType) ) ||
+      (c.chartName === 'rancher-wins-upgrader' && clusterProvider === 'rke2')
+    ) {
+      return false;
+    }
+
+    if (compatibleVersionsFor(c, operatingSystems, showPrerelease).length <= 0) {
+      // There's no versions compatible with the specified os
+      return false;
+    }
+
+    if ( category && !c.categories.includes(category) ) {
+      // The category filter doesn't match
+      return false;
+    }
+
+    if ( searchQuery ) {
+      // The search filter doesn't match
+      const searchTokens = searchQuery.split(/\s*[, ]\s*/).map((x) => ensureRegex(x, false));
+
+      for ( const token of searchTokens ) {
+        const chartDescription = c.chartDescription || '';
+
+        if ( !c.chartNameDisplay.match(token) && !chartDescription.match(token) ) {
+          return false;
+        }
+      }
+    }
+
+    return true;
+  });
+
+  return sortBy(out, ['certifiedSort', 'repoName', 'chartNameDisplay']);
 }
