+++ shell/utils/version.js
@@ -21,10 +21,6 @@ export function sortable(str) {
 }
 
 export function compare(in1, in2) {
-  if (in1 === undefined && in2 === undefined) {
-    return null;
-  }
-
   if ( !in1 ) {
     return 1;
   }
